+++ inst/app/server.R
@@ -286,18 +286,10 @@ server <- function(input, output, session) {
     taxa <- unique(rv$station_summary[, c("name", "AphiaID")])
     taxa <- taxa[!is.na(taxa$name), ]
     
-    # Assign plankton groups using additional custom grouping
-    custom_groups <- list(
-      Cryptophytes = list(phylum = "Cryptophyta"),
-      `Mesodinium spp.` = list(genus = "Mesodinium")
-    )
-    
     groups <- tryCatch(
-      SHARK4R::assign_phytoplankton_group(
+      algaware::assign_phyto_groups(
         scientific_names = taxa$name,
-        aphia_ids        = taxa$AphiaID,
-        custom_groups    = custom_groups,
-        verbose          = FALSE
+        aphia_ids        = taxa$AphiaID
       ),
       error = function(e) rep("Other", nrow(taxa))
     )